cała lista recognised pattern z prawdopodobieństwem (jeszcze bez obrazków); zmienione probability w RecognisedPatterns na float

+++ PatternRecogniser/Controllers/TemporeryRequestsController.cs
@@ -165,7 +165,7 @@ namespace PatternRecogniser.Controllers
                 RecognisedPatterns rp = new RecognisedPatterns()
                 {
                     pattern = paterns.ElementAt(r.Next(maxPaterns)),
-                    probability = r.NextDouble()
+                    probability = (float) r.NextDouble()
                 };
                 ps.Add(rp);
             }
